requirements.txt
fastapi
uvicorn
python-multipart
Pillow
pydantic
python-dotenv

# OCR Engine (CPU-friendly)
easyocr

# PDF Processing
PyMuPDF

# Image Processing
numpy
